Fixes precision rounding for Score Rank percentage
The current implementation only works for settings.DecimalPrecision = 2. For settings.DecimalPrecision = 1, it will always round off and truncate the tenths place since the roundMultiple = 100. For settings.DecimalPrecision > 2, it will include too many decimal places when rounding down, since roundMultiple is increasing in the order of 100x, not 10x. The latter isn't noticed because ToString will truncate the extra decimal places. For the former, it is noticed because the tenths decimal place never updates.